jan_10_26_presentation_Agentic RAG Demo/requirements.txt
streamlit>=1.32.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.4.0
ragas>=0.1.0
pypdf>=4.0.0
python-dotenv>=1.0.0
openai>=1.0.0
tiktoken>=0.5.0
